Refactor configuration parsing to an ``ArgumentsManager`` class
