+++ pydocstringformatter/configuration/__init__.py
@@ -1,11 +1,5 @@
-from pydocstringformatter.configuration.argument_parsing import (
-    _parse_options,
-    _register_arguments,
-    _register_arguments_formatters,
-)
+from pydocstringformatter.configuration.arguments_manager import ArgumentsManager
 
 __all__ = [
-    "_register_arguments",
-    "_register_arguments_formatters",
-    "_parse_options",
+    "ArgumentsManager",
 ]
